requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
openai==1.3.0
joblib==1.3.2
pandas==2.1.4 